logger_utils.py
import logging
from opencensus.ext.azure.log_exporter import AzureLogHandler

logger = logging.getLogger(__name__)
logger.addHandler(AzureLogHandler(connection_string='InstrumentationKey=56b84e5f-199f-4450-a21d-19f7678d797c;IngestionEndpoint=https://southindia-0.in.applicationinsights.azure.com/;LiveEndpoint=https://southindia.livediagnostics.monitor.azure.com/'))
for num in range(5):
    logger.warning(f"Log Entry - {num}")