requirements.txt
absl-py==1.4.0
altair==5.0.1
appdirs==1.4.4
astunparse==1.6.3
attrs==23.1.0
azure-core==1.28.0
azure-identity==1.13.0
azure-storage-blob==12.17.0
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.4
cloudpickle==2.2.1
colorama==0.4.6
contourpy==1.1.0
cryptography==41.0.2
cssselect==1.2.0
cycler==0.11.0
dask==2023.7.0
decorator==5.1.1
dm-tree==0.1.8
fake-useragent==1.1.3
feedparser==6.0.10
filelock==3.12.2
Flask==2.3.2
Flask-Cors==4.0.0
flatbuffers==23.5.26
fonttools==4.40.0
frozendict==2.3.8
fsspec==2023.6.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.32
google-auth==2.21.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.56.0
h5py==3.9.0
html5lib==1.1
idna==3.4
importlib-metadata==6.8.0
investpy==1.0.8
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.1
jsonschema==4.18.2
jsonschema-specifications==2023.6.1
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.0
locket==1.0.0
lxml==4.9.3
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
mdurl==0.1.2
mpmath==1.3.0
msal==1.22.0
msal-extensions==1.0.0
multitasking==0.0.11
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
parse==1.19.1
partd==1.4.0
Pillow==9.5.0
portalocker==2.7.0
protobuf==4.23.3
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydeck==0.8.1b0
pyee==8.2.2
Pygments==2.15.1
PyJWT==2.8.0
Pympler==1.0.1
pyparsing==3.0.9
pyppeteer==1.0.2
pyquery==2.0.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0
referencing==0.29.1
requests==2.31.0
requests-html==0.10.0
requests-oauthlib==1.3.1
rich==13.4.2
rpds-py==0.8.10
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.0
soupsieve==2.4.1
streamlit==1.24.1
sympy==1.12
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
tensorflow-probability==0.20.1
termcolor==2.3.0
threadpoolctl==3.1.0
toml==0.10.2
toolz==0.12.0
torch==2.0.1
tornado==6.3.2
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3.1
Unidecode==1.3.6
urllib3==1.26.16
validators==0.20.0
w3lib==2.1.1
watchdog==3.0.0
webencodings==0.5.1
websockets==10.4
Werkzeug==2.3.6
wrapt==1.15.0
yahoo-fin==0.8.9.1
yfinance==0.2.22
zipp==3.15.0
azure-appconfiguration==1.4.0
applicationinsights==0.11.10
opencensus-ext-azure==1.1.9

